Move OrganizationMembersService tests to a single file
